git-svn-id: http://10.72.141.11:8000/svn/ryuz/software/NekoChat/trunk@272 a23ed8bb-8edb-4d30-a0ab-10e518164151

+++ NekoChat/MainForm.cs
@@ -163,8 +163,8 @@ namespace NekoChat
 
             Microsoft.Win32.RegistryKey regkey3 = Microsoft.Win32.Registry.CurrentUser.CreateSubKey(@"Software\Ryuz\NekoChat\options\");
             regkey3.SetValue("keywords", this.keywords);
-            regkey3.SetValue("keyword-window-popup", this.keywordWindowPopup ? 1 : 0);
-            regkey3.SetValue("keyword-taskbar-blink", this.keywordTaskbarBlink ? 1 : 0);
+            regkey3.SetValue("keyword-window-popup", (int)(this.keywordWindowPopup ? 1 : 0));
+            regkey3.SetValue("keyword-taskbar-blink", (int)(this.keywordTaskbarBlink ? 1 : 0));
             regkey3.SetValue("password", CryptString.EncryptString(Program.Password, registryKey));
             regkey3.Close();
         }
